Started to work on growthbook-kotlinx-serialization module
