+++ Core/src/commonMain/kotlin/com/sdk/growthbook/model/GBValues.kt
@@ -1,0 +1,28 @@
+package com.sdk.growthbook.model
+
+data class GBBoolean(val value: Boolean): GBValue()
+data class GBString(val value: String): GBValue()
+data class GBNumber(val value: Number): GBValue()
+class GBArray(
+    value: List<GBValue>
+): GBValue(), List<GBValue> by value
+data class GBJson(
+    private val value: Map<String, GBValue>,
+): GBValue(), Map<String, GBValue> by value
+
+fun Number.toGbNumber() = GBNumber(this)
+fun Boolean.toGbBoolean() = GBBoolean(this)
+
+sealed class GBValue {
+    data object Unknown: GBValue()
+
+    companion object {
+        internal fun from(anyObj: Any): GBValue =
+            when(anyObj) {
+                is Boolean -> GBBoolean(anyObj)
+                is String -> GBString(anyObj)
+                is Number -> GBNumber(anyObj)
+                else -> Unknown
+            }
+    }
+}
