+++ GrowthBookKotlinxSerialization/build.gradle.kts
@@ -5,18 +5,13 @@ plugins {
     kotlin("multiplatform")
     id("com.android.library")
     kotlin("plugin.serialization")
-    id("org.jetbrains.dokka") version "1.9.10"
+    id("org.jetbrains.dokka") version "1.8.10"
 }
 
 group = "io.growthbook.sdk"
-version = "4.0.0-alpha"
+version = "1.0.0"
 
 kotlin {
-
-    val ktorVersion = "2.1.2"
-    val kryptoVersion = "2.7.0"
-    val kotlinxSerializationVersion = libs.versions.kotlinxSerialization.get()
-
     androidTarget {
         publishLibraryVariants("release")
     }
@@ -47,88 +42,28 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
-                implementation("org.jetbrains.kotlin:kotlin-stdlib-common")
-                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.0")
-                implementation("com.ionspin.kotlin:bignum:0.3.3")
-                implementation("com.soywiz.korlibs.krypto:krypto:$kryptoVersion")
-
-                implementation(
+                val kotlinxSerializationVersion = libs.versions.kotlinxSerialization.get()
+                api(
                     "org.jetbrains.kotlinx:kotlinx-serialization-json:$kotlinxSerializationVersion"
                 )
-                api(project(":Core"))
-                implementation(project(":GrowthBookKotlinxSerialization"))
+                implementation(project(":Core"))
             }
         }
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test-common"))
-                implementation(kotlin("test-annotations-common"))
-                implementation(
-                    "org.jetbrains.kotlinx:kotlinx-serialization-json:$serializationVersion"
-                )
-                implementation("io.ktor:ktor-serialization-kotlinx-json:$ktorVersion")
-            }
-        }
-        val androidMain by getting {
-            dependencies {
-                implementation("androidx.startup:startup-runtime:1.1.1")
-                implementation("com.soywiz.korlibs.krypto:krypto-android:$kryptoVersion")
-            }
-        }
-        val androidUnitTest by getting {
-            dependencies {
-                implementation(kotlin("test-junit"))
-                implementation("junit:junit:4.13.2")
-                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-test")
-                implementation("io.ktor:ktor-client-mock:$ktorVersion")
-            }
-        }
-
-        val jvmMain by getting {
-            dependencies {
-                implementation("io.ktor:ktor-client-java:$ktorVersion")
-                implementation("com.soywiz.korlibs.krypto:krypto-jvm:$kryptoVersion")
-            }
-        }
-        val jvmTest by getting {
-            dependencies {
-                implementation ("org.jetbrains.kotlin:kotlin-test-junit")
-                implementation("com.soywiz.korlibs.krypto:krypto-jvm:$kryptoVersion")
-
-                // https://mvnrepository.com/artifact/org.jetbrains.kotlinx/kotlinx-coroutines-test
-                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.9.0")
-
-                implementation("io.mockk:mockk:1.13.16")
-            }
-        }
-
     }
-
 }
 
 android {
     compileSdk = 34
-    namespace = "com.sdk.growthbook"
-    sourceSets["main"].manifest.srcFile("src/androidMain/AndroidManifest.xml")
+    namespace = "com.sdk.growthbook.kotlinx.serialization"
     defaultConfig {
         minSdk = 21
-
-        consumerProguardFiles("consumer-rules.pro")
     }
     buildTypes {
         release {
             isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
         }
         debug {
             isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
         }
     }
     compileOptions {
@@ -184,7 +119,7 @@ publishing {
             artifact(javadocJar)
             pom {
                 name.set("kotlin")
-                description.set("Powerful A/B testing SDK for Kotlin")
+                description.set("Core module of GrowthBook Kotlin SDK")
                 licenses {
                     license {
                         name.set("MIT")
@@ -202,8 +137,8 @@ publishing {
                 }
                 developers {
                     developer {
-                        name.set("Nicholas Pearson")
-                        email.set("nicholaspearson918@gmail.com")
+                        name.set("Bohdan Kim")
+                        email.set("user576g@gmail.com")
                     }
                 }
             }
