+++ gradle/libs.versions.toml
@@ -1,0 +1,2 @@
+[versions]
+kotlinxSerialization = "1.3.3"
