+++ settings.gradle.kts
@@ -11,3 +11,4 @@ include(":GrowthBook")
 include(":Core")
 include(":NetworkDispatcherKtor")
 include(":NetworkDispatcherOkHttp")
+include(":GrowthBookKotlinxSerialization")
